2022.09.14

+++ product/src/main/java/com/kh/product/domain/Product.java
@@ -11,5 +11,5 @@ public class Product {
   private Long pid;       //상품번호  pid number(10),
   private String pname;   //상품명  pname varchar(30),
   private Integer count;  //상품수량  count number(10),
-  private Integer price;  //상품가격  price number(10)
+  private Long price;  //상품가격  price number(10)
 }
